[coredump] Refactor info and dump collection
Refactor the plugin to adjust how `coredumpctl info` and coredump file
collection is handled.

Plugin will now collect the compressed coredump file for coredump
entries for which coredumpctl reports that the file is present, for the
first X number of coredump files where X is the new `dumps` plugin
option.

A second new plugin option, `executable`, can be used to specific a
regex string to match coredump entries against to determine if they
should be collected or not. The default behavior is to collect for all
entries.

A symlink for coredump file collections will be dropped in the plugin
directory to aide in associating dump files with coredumpctl entries, as
it may not always be obvious for end users based on filenames alone.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>